Fields editor and edited moved from WorkspaceMaterial to HtmlMaterial
