+++ muikku-core-plugins/src/main/java/fi/otavanopisto/muikku/plugins/material/rest/HtmlRestMaterial.java
@@ -1,5 +1,7 @@
 package fi.otavanopisto.muikku.plugins.material.rest;
 
+import java.util.Date;
+
 import fi.otavanopisto.muikku.plugins.material.model.MaterialViewRestrict;
 
 public class HtmlRestMaterial extends RestMaterial {
@@ -7,10 +9,12 @@ public class HtmlRestMaterial extends RestMaterial {
   public HtmlRestMaterial() {
   }
   
-  public HtmlRestMaterial(Long id, String title, String contentType, String html, String license, MaterialViewRestrict visibility) {
+  public HtmlRestMaterial(Long id, String title, String contentType, String html, String license, MaterialViewRestrict visibility, String editor, Date edited) {
     super(id, title, license, visibility);
     this.contentType = contentType;
     this.html = html;
+    this.editor = editor;
+    this.edited = edited;
   }
   
   public String getContentType() {
@@ -28,8 +32,26 @@ public class HtmlRestMaterial extends RestMaterial {
   public void setHtml(String html) {
     this.html = html;
   }
-  
+
+  public String getEditor() {
+    return editor;
+  }
+
+  public void setEditor(String editor) {
+    this.editor = editor;
+  }
+
+  public Date getEdited() {
+    return edited;
+  }
+
+  public void setEdited(Date edited) {
+    this.edited = edited;
+  }
+
   private String html;
   private String contentType;
+  private String editor;
+  private Date edited;
 
 }
