+++ muikku-core-plugins/src/main/java/fi/otavanopisto/muikku/plugins/material/rest/MaterialRESTService.java
@@ -1,7 +1,6 @@
 package fi.otavanopisto.muikku.plugins.material.rest;
 
 import java.util.ArrayList;
-import java.util.Date;
 import java.util.List;
 
 import javax.ejb.Stateful;
@@ -19,8 +18,6 @@ import javax.ws.rs.core.Response.Status;
 
 import org.apache.commons.lang3.StringUtils;
 
-import fi.otavanopisto.muikku.model.users.EnvironmentRoleArchetype;
-import fi.otavanopisto.muikku.model.users.UserEntity;
 import fi.otavanopisto.muikku.plugin.PluginRESTService;
 import fi.otavanopisto.muikku.plugins.material.MaterialController;
 import fi.otavanopisto.muikku.plugins.material.model.Material;
@@ -29,7 +26,6 @@ import fi.otavanopisto.muikku.plugins.workspace.model.WorkspaceMaterial;
 import fi.otavanopisto.muikku.plugins.workspace.model.WorkspaceNode;
 import fi.otavanopisto.muikku.security.MuikkuPermissions;
 import fi.otavanopisto.muikku.session.SessionController;
-import fi.otavanopisto.muikku.users.UserEntityController;
 import fi.otavanopisto.security.rest.RESTPermit;
 import fi.otavanopisto.security.rest.RESTPermit.Handling;
 
@@ -49,9 +45,6 @@ public class MaterialRESTService extends PluginRESTService {
   
   @Inject
   private SessionController sessionController;
-  
-  @Inject
-  private UserEntityController userEntityController;
   
   @GET
 //  @Path("/material/{ID:[0-9]*}/workspaceMaterials/")
@@ -218,28 +211,10 @@ public class MaterialRESTService extends PluginRESTService {
     Long nextSiblingId = workspaceNodeNextSibling != null ? workspaceNodeNextSibling.getId() : null;
     WorkspaceNode workspaceNode = workspaceMaterialController.findWorkspaceNodeById(workspaceMaterial.getId());
     
-    boolean isStaffRole = sessionController.hasAnyRole(
-        EnvironmentRoleArchetype.ADMINISTRATOR,
-        EnvironmentRoleArchetype.STUDY_PROGRAMME_LEADER,
-        EnvironmentRoleArchetype.TEACHER,
-        EnvironmentRoleArchetype.STUDY_GUIDER
-    );
-    
-    String editorName = null;
-    Date edited = null;
-    
-    if (isStaffRole) {
-      if (workspaceMaterial.getEditor() != null) {
-        UserEntity editorUserEntity = userEntityController.findUserEntityById(workspaceMaterial.getEditor());
-        editorName = userEntityController.getName(editorUserEntity, true).getDisplayName();
-      }
-      edited = workspaceMaterial.getEdited();
-    }
-    
     return new fi.otavanopisto.muikku.plugins.workspace.rest.model.WorkspaceMaterial(workspaceMaterial.getId(), workspaceMaterial.getMaterialId(),
         workspaceMaterial.getParent() != null ? workspaceMaterial.getParent().getId() : null, nextSiblingId, workspaceMaterial.getHidden(), 
         workspaceMaterial.getAssignmentType(), workspaceMaterial.getCorrectAnswers(), workspaceMaterial.getPath(), workspaceMaterial.getTitle(),
-        workspaceNode.getLanguage(), workspaceMaterial.getMaxPoints(), workspaceMaterial.getAi(), workspaceMaterial.isExamAssignment(), workspaceMaterial.getExtraInfo(), editorName, edited);
+        workspaceNode.getLanguage(), workspaceMaterial.getMaxPoints(), workspaceMaterial.getAi(), workspaceMaterial.isExamAssignment(), workspaceMaterial.getExtraInfo());
   }
   
 }
