+++ muikku-core-plugins/src/main/java/fi/otavanopisto/muikku/plugins/workspace/dao/WorkspaceMaterialDAO.java
@@ -2,7 +2,6 @@ package fi.otavanopisto.muikku.plugins.workspace.dao;
 
 import java.util.ArrayList;
 import java.util.Collections;
-import java.util.Date;
 import java.util.List;
 import java.util.Set;
 
@@ -236,11 +235,4 @@ public class WorkspaceMaterialDAO extends CorePluginsDAO<WorkspaceMaterial> {
     workspaceMaterial.setExtraInfo(extraInfo);
     return persist(workspaceMaterial);
   }
-  
-  public WorkspaceMaterial updateEditorAndEdited(WorkspaceMaterial workspaceMaterial, Long editor, Date edited) {
-    workspaceMaterial.setEditor(editor);
-    workspaceMaterial.setEdited(edited);
-    return persist(workspaceMaterial);
-  }
-
 }
