+++ updates/1.185/260907_workspace_material_extra_info.xml
@@ -5,6 +5,6 @@
      alter table WorkspaceMaterial add column extraInfo varchar(255);
    </sql>
    <sql>
-     alter table WorkspaceMaterial add column (editor bigint(20), edited datetime);
+     alter table HtmlMaterial add column (editor bigint(20), edited datetime);
    </sql>
 </update>
